Move query related redux to separate file
