+++ public/app/features/explore/ExploreToolbar.tsx
@@ -11,16 +11,7 @@ import { DataQuery, RawTimeRange, TimeRange, TimeZone } from '@grafana/data';
 import { DataSourcePicker } from 'app/core/components/Select/DataSourcePicker';
 import { StoreState } from 'app/types/store';
 import { createAndCopyShortLink } from 'app/core/utils/shortLinks';
-import {
-  cancelQueries,
-  changeDatasource,
-  changeRefreshInterval,
-  clearQueries,
-  runQueries,
-  splitClose,
-  splitOpen,
-  syncTimes,
-} from './state/actions';
+import { changeDatasource, changeRefreshInterval, splitClose, splitOpen, syncTimes } from './state/actions';
 import { updateLocation } from 'app/core/actions';
 import { getTimeZone } from '../profile/state/selectors';
 import { updateTimeZoneForSession } from '../profile/state/reducers';
@@ -33,6 +24,7 @@ import { RunButton } from './RunButton';
 import { LiveTailControls } from './useLiveTailControls';
 import { getExploreDatasources } from './state/selectors';
 import { setDashboardQueriesToUpdateOnLoad } from '../dashboard/state/reducers';
+import { cancelQueries, clearQueries, runQueries } from './state/query';
 
 const { ButtonSelect } = LegacyForms;
