+++ public/app/features/explore/useLiveTailControls.ts
@@ -1,10 +1,11 @@
 import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { changeRefreshInterval, runQueries } from './state/actions';
+import { changeRefreshInterval } from './state/actions';
 import { setPausedStateAction } from './state/actionTypes';
 import { RefreshPicker } from '@grafana/ui';
 import { ExploreId } from '../../types';
+import { runQueries } from './state/query';
 
 /**
  * Hook that gives you all the functions needed to control the live tailing.
